services/intelligence/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
asyncpg==0.29.0
google-cloud-pubsub==2.21.1
google-cloud-aiplatform==1.48.0
google-cloud-storage==2.16.0
google-cloud-bigquery==3.20.1
google-cloud-bigquery-storage==2.25.0
protobuf==4.25.3
grpcio==1.63.0
grpcio-status==1.63.0

# ML & Data
xgboost==2.0.3
scikit-learn==1.4.2
dask[complete]==2024.4.1
pyarrow==15.0.2
pandas==2.2.2
numpy==1.26.4

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
httpx==0.27.0

# Dev / Security
python-dotenv==1.0.1
